Add files via upload
index.html: This HTML file represents a login page with input fields for username, email, and password. It has a background image set using CSS to create a visually appealing layout. The form submits user credentials to the server for authentication.

success.html: This HTML file serves as a success page displayed after a user successfully logs in. It retrieves the username from the URL query parameter and displays a welcome message along with the username. After a brief delay of 2 seconds, it automatically redirects the user to the product catalog page.

app.js: This Node.js application handles the backend logic for the login process. It uses Express.js to define routes for serving static files, handling form submissions, and managing redirections. Upon receiving the login form data, it saves the user's information to a JSON file and redirects the user to the success page with the username as a query parameter.

product-catalog.html: This HTML file represents a product catalog page displaying various electronic products with their descriptions and prices. It utilizes CSS to style the layout and make it visually appealing.

bg_login.jpg: This image file serves as the background image for the login page (index.html). It enhances the visual appearance of the page and provides a better user experience.